+++ changelog.md
@@ -4,7 +4,7 @@
 
 ## Standard library additions and changes
 
-
+- Added `ioutils` module containing `duplicate` and `duplicateTo` to duplicate `FileHandle` using C function `dup` and `dup2`.
 
 ## Language changes
 
@@ -17,4 +17,3 @@
 
 
 ## Tool changes
-